+++ ShopifySharp/Entities/Charge.cs
@@ -35,9 +35,9 @@ namespace ShopifySharp
         /// <summary>
         /// The price of the application charge.
         /// </summary>
-        /// <remarks>Shopify returns this as a string, but JSON.net should be able to convert it to a double.</remarks>
+        /// <remarks>Shopify returns this as a string, but JSON.net should be able to convert it to a decimal.</remarks>
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// The URL the customer is sent to once they accept/decline a charge.
